+++ apicast/cli.py
@@ -7,7 +7,7 @@ from docopt import docopt, DocoptExit
 from tabulate import tabulate
 
 from apicast import __appname__, __version__
-from apicast.core import grok_beeflight_forecast
+from apicast.core import grok_beeflight_forecast, dwd_beeflight_forecast_stations, dwd_beeflight_site_url_by_slug
 from apicast.util import normalize_options, setup_logging
 
 log = logging.getLogger(__name__)
@@ -18,9 +18,9 @@ def run():
     Access bee flight forecast information published by Deutscher Wetterdienst (DWD).
 
     Usage:
-      apicast stations
-      apicast --url=<url>
-      apicast --station=<station>
+      apicast beeflight stations [--site-slugs]
+      apicast beeflight forecast --url=<url>
+      apicast beeflight forecast --station=<station>
       apicast --version
       apicast (-h | --help)
 
@@ -37,7 +37,7 @@ def run():
         apicast stations
 
         # Display bee flight forecast for Potsdam
-        apicast berlin_brandenburg/potsdam
+        apicast --station=berlin_brandenburg/potsdam
 
     """
 
@@ -57,19 +57,38 @@ def run():
     log.debug('Options: {}'.format(json.dumps(options, indent=4)))
 
     # Run command.
-    if options.url:
+    if options.stations:
+        stations = dwd_beeflight_forecast_stations()
 
-        # Fetch and extract forecast information
+        if options.site_slugs:
+            slugs = []
+            for station in stations:
+                for site in station.sites:
+                    slugs.append(site.slug)
+            print(json.dumps(slugs, indent=4))
+        else:
+            print(json.dumps(stations, indent=4))
+
+    # Fetch and extract forecast information.
+    elif options.url:
         result = grok_beeflight_forecast(options.url)
+        format_beeflight_forecast(result)
 
-        data = result['data']
-        if not data:
-            raise ValueError('No data found or unable to parse')
+    elif options.station:
+        url = dwd_beeflight_site_url_by_slug(options.station)
+        result = grok_beeflight_forecast(url)
+        format_beeflight_forecast(result)
 
-        # Report about weather station / observation location
-        print()
-        print(u'### Prognose des Bienenfluges in {}'.format(result['station']))
-        print()
 
-        # Output forecast data
-        print(tabulate(data[1:], headers=data[0], showindex=False, tablefmt='pipe'))
+def format_beeflight_forecast(result):
+    data = result['data']
+    if not data:
+        raise ValueError('No data found or unable to parse')
+
+    # Report about weather station / observation location
+    print()
+    print(u'### Prognose des Bienenfluges in {}'.format(result['station']))
+    print()
+
+    # Output forecast data
+    print(tabulate(data[1:], headers=data[0], showindex=False, tablefmt='pipe'))